Promote MavenWithFallback detector replacing MvnCli (#1756)
* Promote MavenWithFallback detector replacing MvnCli

* Fix and add test

* Small nit

* Update maven markdown

* Update docs/detectors/maven.md


* Mask URL

* Nit

* Nit

---------
