+++ src/Microsoft.ComponentDetection.Detectors/maven/MavenCommandService.cs
@@ -2,7 +2,6 @@
 namespace Microsoft.ComponentDetection.Detectors.Maven;
 
 using System;
-using System.Collections.Concurrent;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
@@ -19,19 +18,6 @@ internal class MavenCommandService : IMavenCommandService
     internal const string MvnVersionArgument = "--version";
 
     internal static readonly string[] AdditionalValidCommands = ["mvn.cmd"];
-
-    /// <summary>
-    /// Per-location semaphores to prevent concurrent Maven CLI executions for the same pom.xml.
-    /// This allows multiple detectors (e.g., MvnCliComponentDetector and MavenWithFallbackDetector)
-    /// to safely share the same output file without race conditions.
-    /// </summary>
-    private readonly ConcurrentDictionary<string, SemaphoreSlim> locationLocks = new();
-
-    /// <summary>
-    /// Tracks locations where dependency generation has completed successfully.
-    /// Used to skip duplicate executions when multiple detectors process the same pom.xml.
-    /// </summary>
-    private readonly ConcurrentDictionary<string, MavenCliResult> completedLocations = new();
 
     private readonly ICommandLineInvocationService commandLineInvocationService;
     private readonly IMavenStyleDependencyGraphParserService parserService;
@@ -58,56 +44,6 @@ internal class MavenCommandService : IMavenCommandService
     }
 
     public async Task<MavenCliResult> GenerateDependenciesFileAsync(ProcessRequest processRequest, CancellationToken cancellationToken = default)
-    {
-        var pomFile = processRequest.ComponentStream;
-        var pomDir = Path.GetDirectoryName(pomFile.Location);
-        var depsFilePath = Path.Combine(pomDir, this.BcdeMvnDependencyFileName);
-
-        // Check the cache before acquiring the semaphore to allow fast-path returns
-        // even when cancellation has been requested.
-        if (this.completedLocations.TryGetValue(pomFile.Location, out var cachedResult)
-            && cachedResult.Success
-            && File.Exists(depsFilePath))
-        {
-            this.logger.LogDebug("{DetectorPrefix}: Skipping duplicate \"dependency:tree\" for {PomFileLocation}, already generated", DetectorLogPrefix, pomFile.Location);
-            return cachedResult;
-        }
-
-        // Use semaphore to prevent concurrent Maven CLI executions for the same pom.xml.
-        // This allows MvnCliComponentDetector and MavenWithFallbackDetector to safely share the output file.
-        var semaphore = this.locationLocks.GetOrAdd(pomFile.Location, _ => new SemaphoreSlim(1, 1));
-
-        await semaphore.WaitAsync(cancellationToken);
-        try
-        {
-            // Re-check the cache after acquiring the semaphore in case another caller
-            // completed while we were waiting.
-            if (this.completedLocations.TryGetValue(pomFile.Location, out cachedResult)
-                && cachedResult.Success
-                && File.Exists(depsFilePath))
-            {
-                this.logger.LogDebug("{DetectorPrefix}: Skipping duplicate \"dependency:tree\" for {PomFileLocation}, already generated", DetectorLogPrefix, pomFile.Location);
-                return cachedResult;
-            }
-
-            var result = await this.GenerateDependenciesFileCoreAsync(processRequest, cancellationToken);
-
-            // Only cache successful results. Failed results should allow retries for transient failures,
-            // and caching them would waste memory since the cache check requires Success == true anyway.
-            if (result.Success)
-            {
-                this.completedLocations[pomFile.Location] = result;
-            }
-
-            return result;
-        }
-        finally
-        {
-            semaphore.Release();
-        }
-    }
-
-    private async Task<MavenCliResult> GenerateDependenciesFileCoreAsync(ProcessRequest processRequest, CancellationToken cancellationToken)
     {
         var cliFileTimeout = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
         var timeoutSeconds = -1;
